requirements.txt
streamlit
google-genai
pillow
google-auth